The test file for the website
This file includes the the added function that checks to make sure that all the polls of choices, if not then it should not be displayed. Note also that previous tests will fail because those tests did not create choices for the polls. So I went back and made sure that all those tests were creating polls with choices